feat: Add sendPurchaseOrderRegistration function to handle purchase order registration
This commit adds a new file, send-purchase-order-registration.ts, which contains the implementation of the sendPurchaseOrderRegistration function. This function is responsible for sending a purchase order registration request to the Unicom API.

The function takes a purchaseOrder object of type UnicomAPIPurchaseOrderRequest as a parameter and returns a Promise<void>. It uses the UnicomAPIPurchaseOrderAdapter to adapt the purchase order data and make the API request.

This feature is part of the PurchaseOrder feature and is necessary for creating and registering purchase orders.